Fetch all products from vuex
